Use whitespace delim instead of fixed width format
Turns out vobs/vfld files aren't written in a fixed-width format (I have
no idea how they are supposed to formatted...).

Also, vobs files don't always include `FI` in the list of variables even
though the data is there. So we handle that explicitly by adding the
column if it is missing.